moved listen_device() to mainloop; switched to epoll
I've now discarded usleep() for the mainloop and moved listen_device() to mainloop. Epoll gives us the opportunity, to have a minimalistic mainloop, which is easy on resources and only is activated on hardware events on USB interface 1. So, the mainloop isn't even activated when you type on your keyboard, but only when using the media-keys and S1 - S30 buttons.
Also, I've cleaned up the code a bit by formatting and renaming some variables and constants.